fix: moved utils and middleware tests into root tests directory
