+++ tests/utils.test.ts
@@ -2,7 +2,7 @@ import { testProp, fc } from '@fast-check/jest';
 import { JSONParser } from '@streamparser/json';
 import * as rpcUtils from '@/utils';
 import 'ix/add/asynciterable-operators/toarray';
-import * as rpcTestUtils from '../utils';
+import * as rpcTestUtils from './utils';
 
 describe('utils tests', () => {
   testProp(
